requirements.txt
pylint==3.3.8
pytest==8.4.2
pytest-cov==7.0.0
ipython==9.6.0
black==25.9.0
pandas
pyodbc
psycopg2-binary
python-dotenv
matplotlib
sqlalchemy
import-ipynb